refactor: Extract useMapbox types and layer configs
Organize useMapbox hook into modular structure:
- types.ts (77 lines): RouteClickInfo, UseMapboxOptions, PlaybackState interfaces
- layerConfigs.ts (206 lines): Mapbox layer paint configurations
- useMapbox.ts (1608 lines): Core hook logic (still tightly coupled)
- index.ts (12 lines): Barrel file with re-exports

Original useMapbox.ts now re-exports from index for backwards compatibility.
All 229 tests pass with no changes to imports.
